refactor(configs): 패키지 분리
- 각자 역할에 맞는 패키지로 이동

Related:
- https://github.com/next-step/spring-subway-service-kakao/pull/25#discussion_r564982511
