+++ src/main/java/subway/configurations/AuthenticationPrincipalConfig.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.configurations;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
